requirements.txt
streamlit
python-dotenv
PyPDF2
langchain
langchain-community
langchain-google-genai
faiss-cpu
